+++ src/backend/config/KsnipConfig.h
@@ -31,6 +31,7 @@
 
 #include "KsnipConfigOptions.h"
 #include "src/common/enum/CaptureModes.h"
+#include "src/common/enum/SaveQualityMode.h"
 #include "src/common/helper/PathHelper.h"
 #include "src/common/enum/UploaderType.h"
 #include "src/common/constants/DefaultValues.h"
@@ -103,6 +104,12 @@ public:
 
 	virtual bool useSingleInstance() const;
 	virtual void setUseSingleInstance(bool enabled);
+
+	virtual SaveQualityMode saveQualityMode() const;
+	virtual void setSaveQualityMode(SaveQualityMode mode);
+
+	virtual int saveQualityFactor() const;
+	virtual void setSaveQualityFactor(int factor);
 
 	// Annotator