+++ src/gui/settingsDialog/SaverSettings.h
@@ -62,6 +62,7 @@ private:
 
 private slots:
 	void chooseSaveDirectory();
+	SaveQualityMode getSaveQualityMode();
 };